change fractal

+++ layers_c/array_printf_fix16.h
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-#define float2fixed(val, fractal_width) (int)(val * (float)((2 << fractal_width) - 1))
+#define float2fixed(val, fractal_width) (int16_t)(val * (float)((2 << fractal_width) - 1))
 #define fixed2float(val, fractal_width) (float)val / (float)((2 << fractal_width) -1)
 
 void array_printf_1D_fix16(uint16_t input_length,
